FitHub/requirements.txt
Django~=2.0.6
django-rest-framework
django-rest-swagger
django-crispy-forms